# fixtures/hub-gold-elysia/src/app.ts
import { Elysia } from "elysia";

// hub-gold-elysia — 20-route Elysia TypeScript dialect (secondary to Express flagship).
// `new Elysia()` + `app.get|post|…` + `ctx.params|query` / IDENT `{ params|query }` bags
// + `ctx.set.status` + object/literal returns.
// Plugins / `.use` / lifecycle / macros stay honest holes (**D6447** — no invented runtime).

const app = new Elysia();

app.get("/health", () => true);
app.get("/ping", () => 42);
app.get("/version", () => 1);
app.get("/ready", () => "ok");
app.get("/count", () => 3);
app.get("/flag", () => "chrysalis");
app.get("/build", () => 2026);
app.get("/tier", () => "gold");

app.get("/meta", () => ({ service: "hub-gold-elysia", version: 1 }));

app.post("/echo", () => ({ echo: true }));

app.get("/items", () => true);
app.get("/items/:id", ({ params: { id } }) => ({ id }));
app.post("/items", (ctx) => {
  ctx.set.status = 201;
  return { created: true };
});
app.get("/search", ({ query: { q = "" } }) => ({ q }));
app.put("/items/:id", (ctx) => ({ updated: true, id: ctx.params.id }));
app.delete("/items/:id", () => true);
app.patch("/items/:id", ({ params: { id } }) => ({ patched: true, id }));
app.get("/users/:userId", (ctx) => ctx.params.userId);
app.get("/stats", () => 3);
app.post("/notify", (ctx) => {
  ctx.set.status = 202;
  return { ok: true };
});

export default app;