# fixtures/hub-gold-openapi-cwl/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "items-mini",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "responses": {
          "200": {
            "description": "service health",
            "content": {
              "application/json": {
                "example": { "ok": true }
              }
            }
          }
        }
      }
    },
    "/items": {
      "get": {
        "operationId": "items_list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": { "count": 0 }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "items_create",
        "parameters": [
          { "name": "authorization", "in": "header", "schema": { "type": "string" } },
          { "name": "session", "in": "cookie", "schema": { "type": "string" } },
          { "name": "x-trace", "in": "cookie", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": { "name": "widget" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": { "created": true }
              }
            }
          }
        }
      }
    },
    "/items/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "get": {
        "operationId": "items_show",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": { "id": "1", "name": "widget" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "items_delete",
        "responses": {
          "204": { "description": "deleted, no content" }
        }
      }
    },
    "/search": {
      "get": {
        "operationId": "search",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string", "default": "" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": { "q": "widget" }
              }
            }
          }
        }
      }
    },
    "/raw": {
      "get": {
        "operationId": "raw_unspecified",
        "responses": {
          "200": {
            "description": "body shape not specified by the contract",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    }
  }
}