# fixtures/hub-gold-quarkus/src/HubResource.java
package hub;

import java.util.Map;
import jakarta.ws.rs.*;
import jakarta.ws.rs.core.Response;

/**
 * hub-gold-quarkus — 20-route Quarkus JAX-RS resource dialect (secondary to Spring hub-flagship-java ST).
 * Quarkus HTTP uses standard JAX-RS (@Path / @GET / @PathParam / Response) with jakarta.ws.rs.*.
 * Route surface only — reuses G10012 JAX-RS peels; no CDI / RESTEasy filters / Panache invent (D6447).
 */
@Path("")
public class HubResource {

  @GET
  @Path("/health")
  public boolean health() {
    return true;
  }

  @GET
  @Path("/ping")
  public int ping() {
    return 42;
  }

  @GET
  @Path("/version")
  public int version() {
    return 1;
  }

  @GET
  @Path("/ready")
  public String ready() {
    return "ok";
  }

  @GET
  @Path("/count")
  public int count() {
    return 3;
  }

  @GET
  @Path("/flag")
  public String flag() {
    return "chrysalis";
  }

  @GET
  @Path("/build")
  public int build() {
    return 2026;
  }

  @GET
  @Path("/tier")
  public String tier() {
    return "gold";
  }

  @GET
  @Path("/meta")
  public Map<String, Object> meta() {
    return Map.of("service", "hub-gold-quarkus", "version", 1);
  }

  @POST
  @Path("/echo")
  public Map<String, Object> echo() {
    return Map.of("echo", true);
  }

  @GET
  @Path("/items")
  public boolean items() {
    return true;
  }

  @GET
  @Path("/items/{id}")
  public Map<String, Object> getItem(@PathParam("id") String id) {
    return Map.of("id", id);
  }

  @POST
  @Path("/items")
  public Response createItem() {
    return Response.status(201).entity(Map.of("created", true)).build();
  }

  @GET
  @Path("/search")
  public Map<String, Object> search(@QueryParam("q") @DefaultValue("") String q) {
    return Map.of("q", q);
  }

  @PUT
  @Path("/items/{id}")
  public Map<String, Object> putItem(@PathParam("id") String id) {
    return Map.of("updated", true, "id", id);
  }

  @DELETE
  @Path("/items/{id}")
  public boolean deleteItem(@PathParam("id") String id) {
    return true;
  }

  @PATCH
  @Path("/items/{id}")
  public Map<String, Object> patchItem(@PathParam("id") String id) {
    return Map.of("patched", true, "id", id);
  }

  @GET
  @Path("/users/{userId}")
  public String getUser(@PathParam("userId") String userId) {
    return userId;
  }

  @GET
  @Path("/stats")
  public int stats() {
    return 3;
  }

  @POST
  @Path("/notify")
  public Response notify() {
    return Response.status(202).entity(Map.of("ok", true)).build();
  }
}